Revise: white footer, Team profiles.
